чуток изменил

+++ Lab3/main_window.py
@@ -8,11 +8,13 @@ from datetime import datetime
 
 from PyQt6.QtWidgets import QApplication, QMainWindow, QFileDialog, QMessageBox, QInputDialog, QTableWidget, \
     QTableWidgetItem, QHeaderView, QAbstractItemView, QMenu, QAbstractScrollArea, QAbstractItemView, \
-    QPushButton, QComboBox, QCheckBox, QSpinBox, QDoubleSpinBox, QLabel, QGroupBox, QGridLayout, QFormLayout \
-    , QRadioButton, QButtonGroup, QSizePolicy, QScrollArea, QFrame, QTabWidget, QStackedWidget, QStackedLayout, \
+    QPushButton, QComboBox, QCheckBox, QSpinBox, QDoubleSpinBox, QLabel, QGroupBox, QGridLayout, QFormLayout, \
+    QRadioButton, QButtonGroup, QSizePolicy, QScrollArea, QFrame, QTabWidget, QStackedWidget, QStackedLayout, \
     QStatusBar, QProgressBar, QToolButton, QToolBar, QStyle, QStyleOption, QStylePainter, QStyleFactory, \
     QStyleOptionProgressBar, QStyleOptionSlider, QStyleOptionToolButton, QStyleOptionButton, QStyleOptionTab, \
-    QStyleOptionTabWidgetFrame, QStyleOptionTabBarBase, QLineEdit, QPlainTextEdit, QDateEdit, QTimeEdit
+    QStyleOptionTabWidgetFrame, QStyleOptionTabBarBase, QLineEdit, QPlainTextEdit, QDateEdit, QTimeEdit, \
+    QDateTimeEdit, QCalendarWidget, QDial, QSlider, QScrollBar, QSplitter, QTreeView, QTreeWidget, QTreeWidgetItem, \
+    QHeaderView, QAbstractItemView, QMenu, QAbstractScrollArea, QAbstractItemView, QTableWidget, QTableWidgetItem
 
 from PyQt6.QtGui import QCursor
 from PyQt6.QtCore import Qt
@@ -25,6 +27,8 @@ from Lab2 import return_4
 class MainWindow(QMainWindow):
     def __init__(self):
         super().__init__()
+        self.weeks_checkbox = None
+        self.years_checkbox = None
         self.weeks = None
         self.years = None
         self.button_annots = None
@@ -61,22 +65,31 @@ class MainWindow(QMainWindow):
             self.button_annots.setEnabled(True)
         self.button_dir.show()
 
-        self.years = QLineEdit('Введите количество лет', self)
-        self.years.move(20, 120)
-        self.years.resize(200, 20)
+        self.years = QLineEdit(self)
+        self.years.setPlaceholderText = 'Введите количество лет'
+        self.years.move(20, 130)
+        self.years.resize(170, 20)
         self.years.show()
 
-        self.weeks = QLineEdit('Введите количество недель', self)
-        self.weeks.move(20, 150)
-        self.weeks.resize(200, 20)
+        self.weeks = QLineEdit(self)
+        self.weeks.setPlaceholderText = 'Введите количество недель'
+        self.weeks.move(20, 160)
+        self.weeks.resize(170, 20)
         self.weeks.show()
 
-    def create_annots(self):
-        # copy path_to_dataset to path_to_annots/dataset.csv.temp
+        self.years_checkbox = QCheckBox(self)
+        self.years_checkbox.move(200, 130)
+        self.years_checkbox.resize(200, 20)
+        self.years_checkbox.show()
 
+        self.weeks_checkbox = QCheckBox(self)
+        self.weeks_checkbox.move(200, 160)
+        self.weeks_checkbox.resize(170, 20)
+        self.weeks_checkbox.show()
+
+    def create_annots(self):
         shutil.copy(self.path_to_dataset, self.path_to_dir + '/dataset.csv.temp')
         self.path_to_annots = self.path_to_dir + '/dataset.csv.temp'
-
         self.button_annots.setText('Файл аннотации создан')
 
     def open_dir(self):
@@ -92,5 +105,5 @@ if __name__ == '__main__':
     window.show()
     app.exec()
 
-    # detele temp file
-    os.remove(window.path_to_annots)
+    if window.path_to_annots != '':
+        os.remove(window.path_to_annots)
